About & Redesign
- Moved webapps & website icons to home page
- Moved description and tech to about page
- Moved around assets as needed

TODOs:
- Stylize home page
- Stylize resume
