+++ about/templates/about.html
@@ -1,0 +1,45 @@
+{% extends "app.html" %} {% load static %} {% block page_head %}
+<link rel="stylesheet" href="{% static 'home/css/home.css' %}" />
+{% endblock %}
+{% block laydo_content_1 %}
+<p>
+    Welcome to Laydo Net! This space on the Internet is primarily to showcase my
+    skills and abilities surrounding software development. It's been built entirely
+    from scratch using a handful of web technologies that I'll accumulate below as time rolls on.
+    <br><br>
+    Generally, everything can be found on the menu and as of now, houses small web applications, sites,
+    and my personal blog that I've crafted over time. Most of the content on this site will be focused
+    on application development and will continually evolve until a point that becomes apparent.
+</p>
+<h2>Technologies & Platforms</h2>
+<div class="tech-plat">
+    <div class="tech-plat-item">
+        <a id="djangoUsed" href="https://www.djangoproject.com/"></a>
+        <label for="djangoUsed">Django</label>
+    </div>
+    <div class="tech-plat-item">
+        <a id="pythonUsed" href="https://www.python.org/"></a>
+        <label for="pythonUsed">Python</label>
+    </div>
+    <div class="tech-plat-item">
+        <a id="typescriptUsed" href="https://www.typescriptlang.org/"></a>
+        <label for="typescriptUsed">Typescript</label>
+    </div>
+    <div class="tech-plat-item">
+        <a id="postgresUsed" href="https://www.postgresql.org/"></a>
+        <label for="postgresUsed">Postgres</label>
+    </div>
+    <div class="tech-plat-item">
+        <a id="githubUsed" href="https://github.com/"></a>
+        <label for="githubUsed">GitHub</label>
+    </div>
+    <div class="tech-plat-item">
+        <a id="doUsed" href="https://www.digitalocean.com/"></a>
+        <label for="doUsed">Digital Ocean</label>
+    </div>
+    <div class="tech-plat-item">
+        <a id="jestUsed" href="https://jestjs.io/"></a>
+        <label for="jestUsed">Jest</label>
+    </div>
+</div>
+{% endblock %}
