+++ about/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from about import views
+
+urlpatterns = [
+    path('', views.about, name='about'),
+]
