+++ laydonet/settings.py
@@ -54,13 +54,12 @@ INSTALLED_APPS = [
     'trackiss',
     'jwst',
     'mtg',
-    'webapps',
-    'websites',
     'blog',
     'users',
     'orbiter',
     'channels',
     'resume',
+    'about',
 ]
 
 MIDDLEWARE = [
