+++ laydonet/templates/base.html
@@ -20,8 +20,7 @@
       <!-- Overlay content -->
       <div class="laydo-overlay-content">
         <a class="nav-link" onclick="redirect(event)" id="">LaydoNet</a>
-        <a class="nav-link" onclick="redirect(event)" id="webapps">Apps</a>
-        <a class="nav-link" onclick="redirect(event)" id="websites">Sites</a>
+        <a class="nav-link" onclick="redirect(event)" id="about">About</a>
         <a class="nav-link" onclick="redirect(event)" id="blog">Blog</a>
         <a class="nav-link" onclick="redirect(event)" id="resume">Resume</a>
         {% if user.is_authenticated %}
