--- static/websites/css/websites.css
@@ -1,49 +1,0 @@
-.site-space {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-evenly;
-    flex-wrap: wrap;
-}
-
-.site-claim {
-    margin: 1rem;
-    padding: 1rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    max-height: 16.5rem;
-}
-
-.site-div {
-    min-height: 15rem;
-    min-width: 10rem;
-    background-color: transparent;
-    box-shadow: rgba(0, 0, 0, 0.25) 0 0.0625rem 0.25rem;
-    border-radius: 5%;
-    cursor: pointer;
-    background-position: inherit;
-    background-size: cover;
-    background-repeat: no-repeat;
-}
-
-.site-title {
-    font-size: 1rem;
-}
-
-#wedding {
-    background-image: url('/static/wedding/img/Aieleen-Landen-Engagement-13.jpg');
-}
-#apollosarcade {
-    background-image: url('/static/websites/img/apollo.png');
-}
-
-@media only screen and (min-width: 37.5rem) {
-    .site-div {
-        min-height: 20rem;
-        min-width: 15rem;
-    }
-
-    .site-title {
-        font-size: 2rem;
-    }
-}
