--- webapps/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from webapps import views
-
-urlpatterns = [
-    path('', views.webapps, name='webapps'),
-]
