--- webapps/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-def webapps(request):
-    return render(request, 'webapps.html', {})
