--- websites/templates/websites.html
@@ -1,16 +1,0 @@
-{% extends "app.html" %} {% load static %} {% block page_head %}
-<link rel="stylesheet" href="{% static 'websites/css/websites.css' %}" />
-{% endblock %} 
-{% block laydo_content_1 %}
-<div class="site-space">
-    <div class="site-claim">
-        <div class="site-div" id="wedding" onclick="redirect(event)"></div>
-        <h3 class="site-title">Wedding</h3>
-    </div>
-    <div class="site-claim">
-        <div class="site-div" id="apollosarcade" onclick="redirectToUrl(event)"></div>
-        <h3 class="site-title">Apollo's Arcade</h3>
-    </div>
-</div>
-{% endblock %} {% block page_scripts %}
-{% endblock %}
