--- websites/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from websites import views
-
-urlpatterns = [
-    path('', views.websites, name='websites'),
-]
