--- websites/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-def websites(request):
-    return render(request, 'websites.html', {})
